fix: [FFM-911] fix metrics not updating when active env changes (#3169)
The metrics tab of the feature flag details page used the environment from
route params. When the active environment changes via the environment
select drop down, it changes a query param, but not the route param. This
means that when changing active environment the metrics tab refreshes using
the previous environment.

This commit removes the environment from the route and instead uses the
environment from the query params. This commit also introduces the
`useActiveEnvironment` hook which provides both the current active
environment from the query params and a utility function
`withActiveEnvironment`, which adds the `activeEnvironment` query param
to a URL string

Additions
- adds `useActiveEnvironment` hook

Changes
- updates `toCFFeatureFlagsDetail` route definition function
- changes `FlagActivation` component to trigger history change when active
  environment changes
- updates failed tests expecting previous route match
- wraps usage of `toCFFeatureFlagsDetail` with `withActiveEnvironment`

Refs FFM-911